requirements.txt
dotenv==0.9.9
Flask==3.1.0
flask-cors==5.0.1
mysql-connector==2.2.9
numpy==2.2.4
python-dotenv==1.1.0
scikit-learn==1.6.1
scipy==1.15.2
openai>=0.27.0
textblob>=0.15.3
konply